DOP-3877: Share Gatsby Cloud webhooks with v1 and v2
