+++ api/controllers/v2/github.ts
@@ -1,29 +1,13 @@
 import * as c from 'config';
-import * as crypto from 'crypto';
 import * as mongodb from 'mongodb';
 import { APIGatewayEvent, APIGatewayProxyResult } from 'aws-lambda';
+import { PushEvent } from '@octokit/webhooks-types';
 
 import { JobRepository } from '../../../src/repositories/jobRepository';
 import { ConsoleLogger } from '../../../src/services/logger';
 import { BranchRepository } from '../../../src/repositories/branchRepository';
 import { EnhancedJob, JobStatus } from '../../../src/entities/job';
-import { PushEvent } from '@octokit/webhooks-types';
-
-// This function will validate your payload from GitHub
-// See docs at https://developer.github.com/webhooks/securing/#validating-payloads-from-github
-function validateJsonWebhook(request: APIGatewayEvent, secret: string) {
-  const expectedSignature =
-    'sha256=' +
-    crypto
-      .createHmac('sha256', secret)
-      .update(request.body ?? '')
-      .digest('hex');
-  const signature = request.headers['X-Hub-Signature-256'];
-  if (signature !== expectedSignature) {
-    return false;
-  }
-  return true;
-}
+import { markBuildArtifactsForDeletion, validateJsonWebhook } from '../../handlers/github';
 
 async function prepGithubPushPayload(
   githubEvent: PushEvent,
@@ -135,3 +119,5 @@ export const TriggerBuild = async (event: APIGatewayEvent): Promise<APIGatewayPr
     body: 'Job Queued',
   };
 };
+
+export const MarkBuildArtifactsForDeletion = markBuildArtifactsForDeletion;
