+++ api/handlers/README
@@ -1,0 +1,4 @@
+# API Handlers
+
+This directory hosts the handler functions that are expected to be shared between
+the regular (v1) Autobuilder lambdas and the enhanced (v2) Autobuilder lambdas.
